requirements.txt
anyio==4.13.0
asgiref==3.11.1
certifi==2026.2.25
Django==6.0.3
django-cors-headers==4.9.0
djangorestframework==3.17.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
python-telegram-bot==22.7
sqlparse==0.5.5
